Further development of the LinearOp class and its derivatives/extensions. In this version a few composite classes were added. In the following description A, B, C, and D are assumed to be linear operators.
LinearOpMatrx - this class allows to construct a matrix of linear
operators: M = [A B; C D];

LinearOpChain - this class allows to construct a chain (product) of
linear operators: P = A*B*C;

LinearOpPlus - this class allows to construct a sum of linear
operators: S = A + B + C;

LinearOP - was modified to accept - (unitary minus), +, and *
operators. The result is a LinearOp, LinearOpPlus, and LinearOpChain
respectively.